+++ src/json-crdt-extensions/peritext/__tests__/Peritext.render-block.spec.ts
@@ -47,8 +47,6 @@ const runInlineSlicesTests = (
       editor.saved.insMarker(['p'], {foo: 'bar'});
       expect(view()).toMatchInlineSnapshot(`
 "<>
-  <0>
-    "" {  }
   <p> { foo = "bar" }
     "abcdefghijklmnopqrstuvwxyz" {  }
 "