+++ src/json-crdt-extensions/peritext/lazy/__tests__/import-export-html.spec.ts
@@ -1,0 +1,14 @@
+import {setupKit} from '../../__tests__/setup';
+import {CommonSliceType} from '../../slice';
+import {fromHtml, toViewRange} from '../import-html';
+
+test('a single paragraph', () => {
+  const {peritext} = setupKit();
+  const html = '<p>Hello world</p>';
+  const peritextMl = fromHtml(html);
+  const rangeView = toViewRange(peritextMl);
+  peritext.editor.import(0, rangeView);
+  peritext.refresh();
+  const json = peritext.blocks.toJson();
+  expect(json).toEqual(['', null, [CommonSliceType.p, null, 'Hello world']]);
+});